@@ -1,5 +1,5 @@
 name: shelf_router
-version: 0.7.4
+version: 0.8.0-nullsafety.0
 description: |
   A convinent request router for the shelf web-framework, with support for
   URL-parameters, nested routers and routers generated from source annotations.
@@ -7,12 +7,13 @@ homepage: https://github.com/google/dart-neats/tree/master/shelf_router
 repository: https://github.com/google/dart-neats.git
 issue_tracker: https://github.com/google/dart-neats/labels/pkg:shelf_router
 dependencies:
-  meta: ^1.1.7
-  shelf: ^0.7.3
-  http_methods: ^1.0.0
+  meta: ^1.3.0-nullsafety
+  shelf: ^1.0.0-nullsafety
+  http_methods: #^1.0.0
+    path: ../http_methods
 dev_dependencies:
-  test: ^1.5.1
-  pedantic: ^1.4.0
-  http: ^0.12.0+1
+  test: ^1.16.0-nullsafety
+  pedantic: ^1.10.0-nullsafety
+  http: ^0.13.0-nullsafety
 environment:
-  sdk: '>=2.3.0 <3.0.0'
+  sdk: '>=2.12.0-0 <3.0.0'
